Fix for own duty blip popping in and out of view
It seems that deleting the blip after creation, even within the same frame, does not hide it from the list of locations on the right side of the pause menu map. As a result the list was difficult to navigate due to the player's own callsign appearing for about half a second every 5 seconds.